Move internal/session into cmd/frontend/internal
Only frontend handles user sessions, so moving it into this package to indicate that properly.

Test plan:

Go compiler doesn't complain.
