+++ cmd/frontend/auth/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//cmd/frontend/globals",
         "//cmd/frontend/internal/app/router",
         "//cmd/frontend/internal/app/ui/router",
+        "//cmd/frontend/internal/auth/session",
         "//cmd/frontend/internal/auth/userpasswd",
         "//internal/actor",
         "//internal/auth",
@@ -30,7 +31,6 @@ go_library(
         "//internal/extsvc",
         "//internal/featureflag",
         "//internal/lazyregexp",
-        "//internal/session",
         "//internal/telemetry",
         "//internal/types",
         "//internal/usagestats",
