+++ cmd/frontend/internal/app/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//cmd/frontend/internal/app/router",
         "//cmd/frontend/internal/app/ui",
         "//cmd/frontend/internal/auth/accessrequest",
+        "//cmd/frontend/internal/auth/session",
         "//cmd/frontend/internal/auth/userpasswd",
         "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/routevar",
@@ -48,7 +49,6 @@ go_library(
         "//internal/gitserver",
         "//internal/httpcli",
         "//internal/otlpenv",
-        "//internal/session",
         "//internal/telemetry",
         "//internal/telemetry/telemetryrecorder",
         "//internal/telemetry/telemetrystore/teestore",
