+++ cmd/frontend/internal/auth/bitbucketcloudoauth/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/auth/oauth",
         "//cmd/frontend/internal/auth/providers",
+        "//cmd/frontend/internal/auth/session",
         "//internal/actor",
         "//internal/collections",
         "//internal/conf",
@@ -28,7 +29,6 @@ go_library(
         "//internal/extsvc/bitbucketcloud",
         "//internal/lazyregexp",
         "//internal/licensing",
-        "//internal/session",
         "//internal/telemetry/telemetryrecorder",
         "//lib/errors",
         "//schema",
