+++ cmd/frontend/internal/auth/userpasswd/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/auth/providers",
+        "//cmd/frontend/internal/auth/session",
         "//internal/actor",
         "//internal/auth",
         "//internal/authz",
@@ -40,7 +41,6 @@ go_library(
         "//internal/lazyregexp",
         "//internal/rcache",
         "//internal/redispool",
-        "//internal/session",
         "//internal/suspiciousnames",
         "//internal/telemetry",
         "//internal/telemetry/telemetrystore/teestore",
@@ -77,12 +77,12 @@ go_test(
     ],
     deps = [
         "//cmd/frontend/backend",
+        "//cmd/frontend/internal/auth/session",
         "//internal/actor",
         "//internal/conf",
         "//internal/database",
         "//internal/database/dbmocks",
         "//internal/rcache",
-        "//internal/session",
         "//internal/telemetry",
         "//internal/telemetry/telemetrytest",
         "//internal/txemail",
