+++ cmd/frontend/internal/cli/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//cmd/frontend/internal/app/assetsutil",
         "//cmd/frontend/internal/app/ui",
         "//cmd/frontend/internal/auth/ipallowlist",
+        "//cmd/frontend/internal/auth/session",
         "//cmd/frontend/internal/auth/userpasswd",
         "//cmd/frontend/internal/authz",
         "//cmd/frontend/internal/bg",
@@ -74,7 +75,6 @@ go_library(
         "//internal/requestclient",
         "//internal/requestinteraction",
         "//internal/service",
-        "//internal/session",
         "//internal/symbols",
         "//internal/sysreq",
         "//internal/trace",
